Migrated GET-LINES to a separate crate.
