--- crates/tools/src/tools/get_lines/mod.rs
@@ -1,3 +1,0 @@
-pub mod cli_utils;
-pub mod get_lines_app;
-pub mod models;
